move date methods to CoronaBaseDateIndexDataFrame
