+++ data_pandas_subclasses/CoronaBaseDateIndex.py
@@ -1,7 +1,13 @@
 import datetime as dt
+from datetime import datetime
+from typing import List, TypeVar
+
 import pandas as pd
+import numpy as np
 
 from data_pandas_subclasses.CoronaBase import CoronaBaseSeries, CoronaBaseDataFrame
+
+TNum = TypeVar('TNum', int, float)
 
 
 class CoronaBaseDateIndexSeries(CoronaBaseSeries):
@@ -42,6 +48,54 @@ class CoronaBaseDateIndexDataFrame(CoronaBaseDataFrame):
             if column in df.columns:
                 df.loc[:, column] = pd.to_datetime(df.loc[:, column])
         return df
+
+    def calculate_7d_moving_mean_for_column(self, column_name: str) -> List[TNum]:
+        return [self._calculate_sum_or_mean_for_period_of_days_and_column_and_specific_day(column_name,
+                                                                                           date,
+                                                                                           days_backwards=3,
+                                                                                           period_in_days=7,
+                                                                                           type="mean")
+                for date
+                in self.index]
+
+    def calculate_sum_7d_to_4d_before_for(self, column_name: str) -> List[TNum]:
+        return [self._calculate_sum_or_mean_for_period_of_days_and_column_and_specific_day(column_name,
+                                                                                           date,
+                                                                                           days_backwards=7,
+                                                                                           period_in_days=4)
+                for date
+                in self.index]
+
+    def calculate_sum_3d_to_0d_before_for(self, column_name: str) -> List[TNum]:
+        return [self._calculate_sum_or_mean_for_period_of_days_and_column_and_specific_day(column_name,
+                                                                                           date,
+                                                                                           days_backwards=3,
+                                                                                           period_in_days=4)
+                for date
+                in self.index]
+
+    def _calculate_sum_or_mean_for_period_of_days_and_column_and_specific_day(self,
+                                                                              column_name: str,
+                                                                              date: datetime,
+                                                                              days_backwards: int,
+                                                                              period_in_days: int,
+                                                                              type: str = "sum") -> TNum:
+        date_range = pd.date_range(date - pd.DateOffset(days_backwards), periods=period_in_days)
+        cases = self.loc[self.index.isin(date_range), column_name]
+        if len(cases) == period_in_days & cases.notna().sum() == period_in_days:
+            if type == "sum":
+                return cases.sum()
+            elif type == "mean":
+                return cases.mean()
+        return np.nan
+
+    def calculate_sum_last_7_days_for_column(self, column_name: str) -> List[TNum]:
+        return [self._calculate_sum_or_mean_for_period_of_days_and_column_and_specific_day(column_name,
+                                                                                           date,
+                                                                                           days_backwards=6,
+                                                                                           period_in_days=7)
+                for date
+                in self.index]
 
     def get_last_date(self) -> dt.datetime:
         return self.index.max()
